move container display logic to another module
